Create a model

Create a model

Usage

base_model(..., .validators_before = NULL, .validators_after = NULL)

Arguments

...

model parameters and their type-check functions

.validators_before

list of validators that run before types are checked

.validators_after

list of validators that run after types are checked

Value

model function

Examples

# Models
my_model <- base_model(
  a = is.integer,
  b = is.integer,
  txt = is.character
)

my_model(a = 1L, b = 2L, txt = "My awesome model")
#> $a
#> [1] 1
#> 
#> $b
#> [1] 2
#> 
#> $txt
#> [1] "My awesome model"
#> 
try(my_model(a = 1, b = 2L, txt = "My awesome model"))
#> Error in raise_type_check_error(k, value, type_check) : 
#>   Type check failed.
#> ℹ field: a
#> ✖ value: 1
#> ✖ test: is.integer

# Validate function args
f <- function(a, b) {
  validate_args(a = is.numeric, b = is.numeric)
  a + b
}

f(4, 5)
#> [1] 9
try(f(4, "5"))
#> Error in raise_type_check_error(k, value, type_check) : 
#>   Type check failed.
#> ℹ field: b
#> ✖ value: "5"
#> ✖ test: is.numeric