Modify a model object

Modify a model object

Usage

model_dump(
  .obj,
  exclude = NULL,
  include = NULL,
  exclude_null = FALSE,
  exclude_na = FALSE,
  camels = FALSE
)

Arguments

.obj

model object

exclude

A set of fields to exclude from the output.

include

A set of fields to include in the output.

exclude_null

Whether to drop items with the value `NULL`.

exclude_na

Whether to drop items with the value `NA`.

camels

Whether to convert all keys to camel case.

Value

list